Reuse recent searches component
